Generate monochrome uint8 data instead of vec4f

+++ src/data.ts
@@ -1,6 +1,6 @@
 import * as wasm from "daicom_preprocessor"
 
-export function generateData(width: number, height: number, depth: number, densityFunction: wasm.GeneratedDataType = wasm.GeneratedDataType.Pillars): Float32Array {
+export function generateData(width: number, height: number, depth: number, densityFunction: wasm.GeneratedDataType = wasm.GeneratedDataType.Pillars): Uint8Array {
     console.log("invoking wasm");
     const result = wasm.generate_data(width, height, depth, densityFunction);
     console.log("invokation result", result);
